daemon: Fix race condition when syncing svcs with k8s
Previously, the "sync-lb-maps-with-k8s-services" controller did the
following:

1. Fetched a list of services from pkg/k8s.k8sServiceCache.
2. Iterated over pkg/service.Service cache and removed those which
   couldn't be found in the former list.

As no lock was held during the (1) and (2) steps, any service added to
service.Service after the (1) step was removed.

Fix this race condition by setting the noUpdateAfterRestore flag for
each service during RestoreServices() (happens before the controller
is scheduled) and unset it in UpsertService(); in SyncWithK8sFinished()
remove those services which still has the flag being set.

The fix is based on an assumption that during the k8s service cache sync
period UpsertService() is going to be called for each alive service.

Signed-off-by: Martynas Pumputis <m@lambda.lt>